requirements.txt
fastapi
uvicorn
python-multipart
pandas
requests
beautifulsoup4
google-play-scraper
app-store-scraper
openai
python-dotenv
boto3
